refactor: welcome screen routes
